# contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Athopia client API",
    "version": "1.0.0",
    "description": "Genererad ur athopia-web/lib/api-schemas.ts. Redigera aldrig manuellt."
  },
  "servers": [
    {
      "url": "https://athopia.se"
    }
  ],
  "paths": {
...
    "/api/feed/hero": {
      "get": {
        "operationId": "HeroResponse",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "summary": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "type": {
                              "type": "string"
                            },
                            "title": {
                              "type": "string"
                            },
                            "subtitle": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "source": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "time": {
                              "type": "string"
                            },
                            "href": {
                              "type": "string"
                            },
                            "newsTag": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "sourceCount": {
                              "anyOf": [
                                {
                                  "type": "integer"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            },
                            "storyClusterId": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "importanceTier": {
                              "anyOf": [
                                {
                                  "type": "string",
                                  "enum": [
                                    "breaking",
                                    "major",
                                    "normal",
                                    "noise"
                                  ]
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "type",
                            "title",
                            "time",
                            "href"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "topNews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "subtitle": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "source": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "time": {
                            "type": "string"
                          },
                          "href": {
                            "type": "string"
                          },
                          "newsTag": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "sourceCount": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "storyClusterId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "importanceTier": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "breaking",
                                  "major",
                                  "normal",
                                  "noise"
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "type",
                          "title",
                          "time",
                          "href"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "summary",
                    "topNews"
                  ],
                  "additionalProperties": false,
                  "$schema": "http://json-schema.org/draft-07/schema#"
                }
              }
            }
          },
          "403": {
            "description": "Otillräcklig plan — servern är auktoritativ.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "const": "plan_required"
                    },
                    "feature": {
                      "type": "string"
                    },
                    "requiredPlan": {
                      "type": "string",
                      "enum": [
                        "free",
                        "pro",
                        "elite"
                      ]
                    },
                    "upgradePath": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "code",
                    "feature",
                    "requiredPlan",
                    "upgradePath"
                  ],
                  "additionalProperties": false,
                  "$schema": "http://json-schema.org/draft-07/schema#"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/match/{fixtureId}": {
      "get": {
        "operationId": "FixtureDetail",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "home_team_id": {
                      "type": "integer"
                    },
                    "away_team_id": {
                      "type": "integer"
                    },
                    "home_team_name": {
                      "type": "string"
                    },
                    "away_team_name": {
                      "type": "string"
                    },
                    "home_score": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "away_score": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "kickoff_at": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "status": {
                      "type": "string"
                    },
                    "home_xg": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "away_xg": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "home_pressure": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "away_pressure": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "home_ppda": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "away_ppda": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "home_possession": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "away_possession": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "home_shots": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "away_shots": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "home_shots_on_target": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "away_shots_on_target": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "home_team_id",
                    "away_team_id",
                    "home_team_name",
                    "away_team_name",
                    "home_score",
                    "away_score",
                    "kickoff_at",
                    "status",
                    "home_xg",
                    "away_xg",
                    "home_pressure",
                    "away_pressure",
                    "home_ppda",
                    "away_ppda",
                    "home_possession",
                    "away_possession",
                    "home_shots",
                    "away_shots",
                    "home_shots_on_target",
                    "away_shots_on_target"
                  ],
                  "additionalProperties": false,
                  "$schema": "http://json-schema.org/draft-07/schema#"
                }
              }
            }
          },
          "403": {
            "description": "Otillräcklig plan — servern är auktoritativ.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "const": "plan_required"
                    },
                    "feature": {
                      "type": "string"
                    },
                    "requiredPlan": {
                      "type": "string",
                      "enum": [
                        "free",
                        "pro",
                        "elite"
                      ]
                    },
                    "upgradePath": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "code",
                    "feature",
                    "requiredPlan",
                    "upgradePath"
                  ],
                  "additionalProperties": false,
                  "$schema": "http://json-schema.org/draft-07/schema#"
                }
              }
            }
          }
        }
      }
    },
...
  }
}